+++ exemple/src/components/custom-comp.tsx
@@ -1,10 +1,12 @@
 
-
+import { useState } from "react";
 
 export default function CustomComp() {
+    const [count, setCount] = useState(0);
     return (
         <div>
-            <h1>Custom Comp1</h1>
+            <h1>Custom Comp {count} ++</h1>
+            <button onClick={() => setCount(count + 1)}>Increment</button>
         </div>
     );
 }