+++ src/react-refresh-transform.ts
@@ -1,8 +1,8 @@
 import {
-	transformAsync,
 	type NodePath,
 	type PluginAPI,
 	type PluginObject,
+	transformAsync,
 } from "@babel/core";
 import * as types from "@babel/types";
 import reactRefreshBabel from "react-refresh/babel";