+++ src/types.d.ts
@@ -14,6 +14,16 @@ declare module "@apply-react/HMR-enabled.ts" {
 	export default HMR_ENABLED;
 }
 
+declare module "@apply-react/fast-refresh-enabled.ts" {
+	const FAST_REFRESH_ENABLED: boolean;
+	export default FAST_REFRESH_ENABLED;
+}
+
+declare module "@apply-react/development-mode.ts" {
+	const IS_DEVELOPMENT: boolean;
+	export default IS_DEVELOPMENT;
+}
+
 declare module "@apply-react/404.tsx" {
 	export default function NotFound(): React.JSX.Element;
 }